requirements.txt
Django==6.0.1
djangorestframework==3.16.1
django-cors-headers==4.9.0
gspread==6.2.1
oauth2client==4.1.3
python-decouple==3.8
Pillow==12.1.0
twilio==9.10.0
requests==2.32.5

